scraper/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.2
cssselect==1.2.0
Django==4.2.3
django-environ==0.10.0
environ==1.0
filelock==3.12.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
packaging==23.1
parsel==1.8.1
Protego==0.2.1
proxyscrape==0.3.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyOpenSSL==23.2.0
python-decouple==3.8
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.9.0
scrapy-proxy-pool==0.1.9
scrapy-user-agents==0.1.1
service-identity==23.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tldextract==3.4.4
Twisted==22.10.0
typing-extensions==4.7.1
ua-parser==0.18.0
urllib3==2.0.4
user-agents==2.2.0
w3lib==2.1.1
zope.interface==6.0
